// src/app/core/models/referenceCommericiale.model.ts
import {Validators} from "@angular/forms";

export class ReferenceCommericiale {
  id?: number;
  company_name?: string;
  contact_full_name?: string;
  contact_phone_number?: string;
  contact_designation?: string;
  contact_email?: string;
  contact_approx_amount?: string;
  reference?: string;
  supplier?: string;
  attachement?: string;
  doc?: string;
  contact_approx_amount_currency?: string;
}